# backend/openapi_doc/specs/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Routine Cloud API v1",
    "description": "Routine Cloud API SDK",
    "version": "v1"
  },
  "paths": {
...
    "/routines/": {
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Create Routine",
        "operationId": "routines_create",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "List Routines",
        "operationId": "routines_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_tasks",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include tasks for each routine",
              "default": false,
              "title": "Include Tasks"
            },
            "description": "Include tasks for each routine"
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutineRead"
                  },
                  "title": "Response Routines List"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "Get Routine",
        "operationId": "routines_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "routines"
        ],
        "summary": "Update Routine",
        "operationId": "routines_update",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "routines"
        ],
        "summary": "Delete Routine",
        "operationId": "routines_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}/tasks": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "List Routine Tasks",
        "operationId": "routines_tasks_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskInRoutineRead"
                  },
                  "title": "Response Routines Tasks List"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Add Task To Routine",
        "operationId": "routines_tasks_add",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineTaskAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskInRoutineRead"
                  },
                  "title": "Response Routines Tasks Add"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}/tasks/{position}": {
      "delete": {
        "tags": [
          "routines"
        ],
        "summary": "Remove Task From Routine",
        "operationId": "routines_tasks_remove",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "position",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskInRoutineRead"
                  },
                  "title": "Response Routines Tasks Remove"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}/shares": {
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Create Routine Share",
        "operationId": "routines_shares_create",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineShareCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineShareRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "List Routine Shares",
        "operationId": "routines_shares_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutineShareRead"
                  },
                  "title": "Response Routines Shares List"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessLevel": {
        "type": "string",
        "enum": [
          "owner",
          "start",
          "read"
        ],
        "title": "AccessLevel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RoutineCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "title": "Description",
            "type": "string",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RoutineCreate"
      },
      "RoutineRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "title": "Description",
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "access_level": {
            "$ref": "#/components/schemas/AccessLevel",
            "nullable": true
          },
          "start_mode": {
            "$ref": "#/components/schemas/StartMode",
            "default": "none"
          },
          "notify_mask": {
            "type": "integer",
            "title": "Notify Mask",
            "default": 0
          },
          "tasks": {
            "title": "Tasks",
            "items": {
              "$ref": "#/components/schemas/TaskInRoutineRead"
            },
            "type": "array",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "Routine"
      },
...
      "RoutineTaskAdd": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "position": {
            "title": "Position",
            "type": "integer",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "RoutineTaskAdd"
      },
      "RoutineUpdate": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "nullable": true
          },
          "description": {
            "title": "Description",
            "type": "string",
            "nullable": true
          },
          "start_mode": {
            "$ref": "#/components/schemas/StartMode",
            "nullable": true
          },
          "notify_mask": {
            "title": "Notify Mask",
            "type": "integer",
            "nullable": true
          }
        },
        "type": "object",
        "title": "RoutineUpdate"
      },
...
      "StartMode": {
        "type": "string",
        "enum": [
          "none",
          "follow_owner",
          "follow_any"
        ],
        "title": "StartMode"
      },
...
      "TaskInRoutineRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon_name": {
            "type": "string",
            "title": "Icon Name"
          },
          "sound": {
            "type": "string",
            "title": "Sound"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "icon_name",
          "sound",
          "duration",
          "position"
        ],
        "title": "TaskInRoutineRead"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}